Rename jsx to js
