+++ normandy/control/static/control/js/components/RecipeContainer.js
