+++ normandy/control/static/control/js/components/RecipeHistory.js
@@ -1,7 +1,7 @@
 import React, { PropTypes as pt } from 'react';
 import { push } from 'react-router-redux';
 import moment from 'moment';
-import composeRecipeContainer from './RecipeContainer.jsx';
+import composeRecipeContainer from './RecipeContainer.js';
 import { makeApiRequest } from '../actions/ControlActions.js';
 
 class RecipeHistory extends React.Component {
