+++ normandy/control/static/control/js/components/form_fields/CheckboxField.js
